requirements.txt
asgiref==3.4.1
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
Markdown==3.3.6
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
